Refactor: 검색 로직을 useSearch Hook으로 분리
